continue dissassembly

+++ inc/defines.h
@@ -146,7 +146,7 @@ typedef char     smode_t[4];
 typedef void    *handle_t;
 typedef void    *unknown_t, **punknown_t;
 
-typedef uint32_t pick_t, *ppick_t;
+typedef uint64_t pick_t, *ppick_t;
 
 typedef int      bool_t;
